+++ Plugin.cs
@@ -50,6 +50,7 @@ namespace ChaoticCorruptions
         public static ConfigEntry<bool> CorruptStartingItems { get; set; }
         public static ConfigEntry<bool> RandomizeStartingDecks { get; set; }
         public static ConfigEntry<bool> CompletelyRandomizeStartingDecks { get; set; }
+        public static ConfigEntry<bool> ChaoticStartingDeck { get; set; }
         public static ConfigEntry<bool> CraftableCorruptions { get; set; }
         public static ConfigEntry<int> CraftableCorruptionsCost { get; set; }
         public static ConfigEntry<bool> PurchaseableCorruptPets { get; set; }
@@ -81,6 +82,7 @@ namespace ChaoticCorruptions
             CorruptStartingItems = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "CorruptStartingItems"), false, new ConfigDescription("Forces all starting items to be corrupted."));
             RandomizeStartingDecks = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "RandomizeStartingDecks"), true, new ConfigDescription("Randomizes starting decks from all craftable cards. If CorruptStartingDecks is active, they are all corrupted"));
             CompletelyRandomizeStartingDecks = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "CompletelyRandomizeStartingDecks"), true, new ConfigDescription("Randomizes starting decks from all available cards for each hero's class. If CorruptStartingDecks is active, they are all corrupted"));
+            ChaoticStartingDeck = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "ChaoticStartingDeck"), false, new ConfigDescription("If true, replaces your starting deck with Chaos (discovers a card). If CorruptStartingDecks is active, replaces your deck with CHAOS (discovers a corrupted card). Overrides all other starting deck options."));
             CraftableCorruptions = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "CraftableCorruptions"), true, new ConfigDescription("Makes corrupted cards craftable"));
             CraftableCorruptionsCost = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "CraftableCorruptionsCost"), 800, new ConfigDescription("The cost added to the regular crafting cost that will be added to the card to craft the corrupted version."));
             // PurchaseableCorruptPets = Config.Bind(new ConfigDefinition(PluginInfo.PLUGIN_NAME, "PurchaseableCorruptPets"), false, new ConfigDescription("Makes corrupted cards craftable"));